refactor structure so that pkg is import/runn-able system-wide
