--- envs/gridworld/__init__.py
@@ -1,2 +1,0 @@
-from .gridexplore import GridExplore
-from .gridpath import GridPath
