--- envs/ml2_python/__init__.py
@@ -1,1 +1,0 @@
-from ml2_python.environment import ML2Python
