+++ marlenv/envs/ml2_python/__init__.py
@@ -1,0 +1,1 @@
+from .environment import ML2Python
