+++ marlenv/envs/ml2_python/graphics.py
@@ -3,8 +3,8 @@ import numpy as np
 import torch
 import pygame
 
-from ml2_python.common import Cell, Direction
-from ml2_python.environment import Action
+from .common import Cell, Direction
+from .environment import Action
 
 
 class Color:
