+++ marlenv/envs/ml2_python/python.py
@@ -1,7 +1,7 @@
 """modified from YuriyGuts/snake-ai-reinforcement"""
 from collections import deque
 
-from ml2_python.common import Direction
+from .common import Direction
 
 
 class Python:
