+++ setup.py
@@ -2,14 +2,14 @@ from os import path
 
 from setuptools import setup
 
-setup(name='ml2_marlenv',
+setup(name='marlenv',
       version='0.0.1',
-      url='https://github.com/taemin410/ml2_marlenv',
-      py_modules=['ml2_marlenv'],
-      author='Tae Min Ha',
-      author_email='taemin410@gmail.com',
+      url='https://github.com/kc-ml2/marlenv',
+      py_modules=['marlenv'],
+      author='Tae Min Ha, Daniel Nam',
+      author_email='taemin410@gmail.com, dwtnam@kc-ml2.com',
       # license=open(path.join(path.abspath(path.dirname(__file__)), 'LICENSE')).read(),
       install_requires=[x.strip() for x in
                         open(path.join(path.abspath(path.dirname(__file__)), 'requirements.txt')).readlines()],
-      python_requires='>=3.5',
+      python_requires='>=3.6',
       )
